fix: Remove TestQueue v1 and update Test Queue e2e tests (#1146)
* Move AssignTesterDropdown

* Removed original components/TestQueue and revised tests

* Additional removal of files

* Rename TestQueue2 -> TestQueue

* Address TODOs on removing Test Queue v1 vs v2 conditionals

* Remove unnecessary TODO
