+++ client/components/AddTestToQueueWithConfirmation/queries.js
@@ -48,3 +48,38 @@ export const EXISTING_TEST_PLAN_REPORTS = gql`
     }
   }
 `;
+
+export const ADD_TEST_QUEUE_MUTATION = gql`
+  mutation AddTestPlanReport(
+    $testPlanVersionId: ID!
+    $atId: ID!
+    $exactAtVersionId: ID
+    $minimumAtVersionId: ID
+    $browserId: ID!
+    $copyResultsFromTestPlanVersionId: ID
+  ) {
+    createTestPlanReport(
+      input: {
+        testPlanVersionId: $testPlanVersionId
+        atId: $atId
+        exactAtVersionId: $exactAtVersionId
+        minimumAtVersionId: $minimumAtVersionId
+        browserId: $browserId
+        copyResultsFromTestPlanVersionId: $copyResultsFromTestPlanVersionId
+      }
+    ) {
+      testPlanReport {
+        id
+        at {
+          id
+        }
+        browser {
+          id
+        }
+      }
+      testPlanVersion {
+        id
+      }
+    }
+  }
+`;
