+++ client/components/TestQueueCompletionStatusListItem/BotTestCompletionStatus/index.js
@@ -2,12 +2,7 @@ import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { useTestPlanRunValidatedAssertionCounts } from '../../../hooks/useTestPlanRunValidatedAssertionCounts';
 
-const BotTestCompletionStatus = ({
-  testPlanRun,
-  id,
-  runnableTestsLength,
-  fromTestQueueV2 = false // TODO: Remove when Test Queue v1 is removed
-}) => {
+const BotTestCompletionStatus = ({ testPlanRun, id, runnableTestsLength }) => {
   const {
     totalValidatedAssertions,
     totalPossibleAssertions,
@@ -23,25 +18,14 @@ const BotTestCompletionStatus = ({
 
   return (
     <>
-      {fromTestQueueV2 ? (
-        <ul id={id} style={{ fontSize: '0.875rem' }}>
-          <li>
-            <em>{`Responses for ${testResultsLength} of ${runnableTestsLength} tests recorded`}</em>
-          </li>
-          <li>
-            <em>{`Verdicts for ${totalValidatedAssertions} of ${totalPossibleAssertions} assertions assigned`}</em>
-          </li>
-        </ul>
-      ) : (
-        <ul id={id} className="text-secondary">
-          <li>
-            {`Responses for ${testResultsLength} of ${runnableTestsLength} tests recorded`}
-          </li>
-          <li>
-            {`Verdicts for ${totalValidatedAssertions} of ${totalPossibleAssertions} assertions assigned`}
-          </li>
-        </ul>
-      )}
+      <ul id={id} style={{ fontSize: '0.875rem' }}>
+        <li>
+          <em>{`Responses for ${testResultsLength} of ${runnableTestsLength} tests recorded`}</em>
+        </li>
+        <li>
+          <em>{`Verdicts for ${totalValidatedAssertions} of ${totalPossibleAssertions} assertions assigned`}</em>
+        </li>
+      </ul>
     </>
   );
 };
@@ -64,8 +48,7 @@ BotTestCompletionStatus.propTypes = {
     )
   }).isRequired,
   id: PropTypes.string.isRequired,
-  runnableTestsLength: PropTypes.number.isRequired,
-  fromTestQueueV2: PropTypes.bool
+  runnableTestsLength: PropTypes.number.isRequired
 };
 
 export default BotTestCompletionStatus;
