+++ client/routes/index.js
@@ -8,7 +8,6 @@ import { Reports, Report } from '@components/Reports';
 import CandidateReview from '@components/CandidateReview';
 import SignupInstructions from '@components/SignupInstructions';
 import TestQueue from '@components/TestQueue';
-import TestQueue2 from '@components/TestQueue2';
 import TestRun from '@components/TestRun';
 import UserSettings from '@components/UserSettings';
 import CandidateTestPlanRun from '@components/CandidateReview/CandidateTestPlanRun';
@@ -38,9 +37,7 @@ export default () => (
         </ConfirmAuth>
       }
     />
-    {/* TODO: Deprecate and remove */}
-    <Route exact path="/test-queue-old" element={<TestQueue />} />
-    <Route exact path="/test-queue" element={<TestQueue2 />} />
+    <Route exact path="/test-queue" element={<TestQueue />} />
     <Route
       exact
       path="/test-plan-report/:testPlanReportId"
