+++ client/tests/TestQueue.test.jsx
@@ -3,7 +3,7 @@
  */
 
 import React from 'react';
-import { render, screen, waitFor, fireEvent } from '@testing-library/react';
+import { render, waitFor, fireEvent } from '@testing-library/react';
 import { InMemoryCache } from '@apollo/client';
 import { MockedProvider } from '@apollo/client/testing';
 import { BrowserRouter } from 'react-router-dom';
@@ -15,8 +15,6 @@ import TestQueue from '../components/TestQueue';
 import {
   TEST_QUEUE_PAGE_ADMIN_NOT_POPULATED_MOCK_DATA,
   TEST_QUEUE_PAGE_TESTER_NOT_POPULATED_MOCK_DATA,
-  TEST_QUEUE_PAGE_ADMIN_POPULATED_MOCK_DATA,
-  TEST_QUEUE_PAGE_TESTER_POPULATED_MOCK_DATA,
   TEST_QUEUE_PAGE_BASE_MOCK_DATA
 } from './__mocks__/GraphQLMocks';
 import { AriaLiveRegionProvider } from '../components/providers/AriaLiveRegionProvider';
@@ -40,140 +38,14 @@ const setup = (mocks = []) => {
 describe('Render TestQueue/index.jsx', () => {
   let wrapper;
 
-  describe('[NOT ADMIN] when no test plan reports exist', () => {
+  describe('not admin when no test plan reports exist', () => {
     beforeEach(() => {
       wrapper = setup(TEST_QUEUE_PAGE_TESTER_NOT_POPULATED_MOCK_DATA);
     });
 
     it('renders loading state on initialization', async () => {
-      const { getByTestId } = wrapper;
-      const element = getByTestId('page-status');
-
-      expect(element).toBeTruthy();
-      expect(element).toHaveTextContent('Loading');
-    });
-
-    // TODO: Revise timeout with pageReady check
-    it.skip('renders Test Queue page instructions', async () => {
-      // allow page time to load
-
-      await waitFor(() => new Promise(res => setTimeout(res, 0)));
-
-      const { queryByTestId, getByTestId } = wrapper;
-      const loadingElement = queryByTestId('page-status');
-      const element = getByTestId('test-queue-no-test-plans-p');
-
-      expect(loadingElement).not.toBeInTheDocument();
-      expect(element).toBeTruthy();
-      expect(element).toHaveTextContent(
-        /Please configure your preferred Assistive Technologies in/i
-      );
-    });
-
-    it('renders no AT-specific sections', async () => {
-      // allow page time to load
-      await waitFor(() => new Promise(res => setTimeout(res, 0)));
-
-      const { queryAllByText } = wrapper;
-      const nvdaElements = queryAllByText(/nvda/i);
-      const jawsElements = queryAllByText(/jaws/i);
-      const voiceOverElements = queryAllByText(/voiceover/i);
-
-      expect(nvdaElements.length).toEqual(0);
-      expect(jawsElements.length).toEqual(0);
-      expect(voiceOverElements.length).toEqual(0);
-    });
-
-    it('does not render add test plan modal button', async () => {
-      // allow page time to load
-      await waitFor(() => new Promise(res => setTimeout(res, 0)));
-
       const { queryByTestId } = wrapper;
-      const button = queryByTestId('test-queue-add-test-plan-to-queue-button');
-
-      expect(button).not.toBeTruthy();
-    });
-  });
-
-  describe('[NOT ADMIN] when test plan reports exist', () => {
-    beforeEach(() => {
-      wrapper = setup(TEST_QUEUE_PAGE_TESTER_POPULATED_MOCK_DATA);
-    });
-
-    it('renders loading state on initialization', async () => {
-      const { getByTestId } = wrapper;
-      const element = getByTestId('page-status');
-
-      expect(element).toBeTruthy();
-      expect(element).toHaveTextContent('Loading');
-    });
-
-    it.skip('renders Test Queue page instructions', async () => {
-      // allow page time to load
-      await waitFor(() => new Promise(res => setTimeout(res, 0)));
-
-      const { queryByTestId, getByTestId } = wrapper;
-      const loadingElement = queryByTestId('page-status');
-      const element = getByTestId('test-queue-instructions');
-
-      expect(loadingElement).not.toBeInTheDocument();
-      expect(element).toBeTruthy();
-      expect(element).toHaveTextContent(
-        'Assign yourself a test plan or start executing one that is already assigned to you.'
-      );
-    });
-
-    it('renders AT-specific sections', async () => {
-      // allow page time to load
-      await waitFor(() => new Promise(res => setTimeout(res, 0)));
-
-      const { queryAllByText } = wrapper;
-      const nvdaElements = queryAllByText(/nvda/i);
-      const jawsElements = queryAllByText(/jaws/i);
-      const voiceOverElements = queryAllByText(/voiceover/i);
-
-      expect(nvdaElements.length).toBeGreaterThanOrEqual(1);
-      expect(jawsElements.length).toBeGreaterThanOrEqual(1);
-      expect(voiceOverElements.length).toBeGreaterThanOrEqual(1);
-    });
-
-    it('renders testers are assigned to Test Plans', async () => {
-      // allow page time to load
-      await waitFor(() => new Promise(res => setTimeout(res, 0)));
-
-      const { queryAllByText } = wrapper;
-      const userAAssignedElements = queryAllByText(/foo-bar/i);
-      const userBAssignedElements = queryAllByText(/bar-foo/i);
-      const userCAssignedElements = queryAllByText(/boo-far/i);
-      const assignedTestsElements = queryAllByText(
-        /\d+ of \d+ tests complete/i
-      );
-
-      expect(userAAssignedElements.length).toBeGreaterThanOrEqual(1);
-      expect(userBAssignedElements.length).toBeGreaterThanOrEqual(1);
-      expect(userCAssignedElements.length).toBeGreaterThanOrEqual(1);
-      expect(assignedTestsElements.length).toBeGreaterThanOrEqual(1);
-    });
-
-    it('does not render add test plan modal button', async () => {
-      // allow page time to load
-      await waitFor(() => new Promise(res => setTimeout(res, 0)));
-
-      const { queryByTestId } = wrapper;
-      const button = queryByTestId('test-queue-add-test-plan-to-queue-button');
-
-      expect(button).not.toBeTruthy();
-    });
-  });
-
-  describe('[IS ADMIN] when no test plan reports exist', () => {
-    beforeEach(() => {
-      wrapper = setup(TEST_QUEUE_PAGE_ADMIN_NOT_POPULATED_MOCK_DATA);
-    });
-
-    it('renders loading state on initialization', async () => {
-      const { getByTestId } = wrapper;
-      const element = getByTestId('page-status');
+      const element = queryByTestId('page-status');
 
       expect(element).toBeTruthy();
       expect(element).toHaveTextContent('Loading');
@@ -183,13 +55,19 @@ describe('Render TestQueue/index.jsx', () => {
       // allow page time to load
       await waitFor(() => new Promise(res => setTimeout(res, 0)));
 
-      const { queryByTestId, getByTestId } = wrapper;
+      const { queryByTestId } = wrapper;
       const loadingElement = queryByTestId('page-status');
-      const element = getByTestId('test-queue-no-test-plans-p');
+      const noTestPlansMessageElement = queryByTestId('no-test-plans');
+      const addTestPlanToQueueMessageElement = queryByTestId(
+        'add-test-plans-queue'
+      );
 
       expect(loadingElement).not.toBeInTheDocument();
-      expect(element).toBeTruthy();
-      expect(element).toHaveTextContent(/Add a Test Plan to the Queue/i);
+      expect(noTestPlansMessageElement).toBeTruthy();
+      expect(noTestPlansMessageElement).toHaveTextContent(
+        /There are currently no test plan reports available/i
+      );
+      expect(addTestPlanToQueueMessageElement).not.toBeTruthy();
     });
 
     it('renders no AT-specific sections', async () => {
@@ -206,57 +84,55 @@ describe('Render TestQueue/index.jsx', () => {
       expect(voiceOverElements.length).toEqual(0);
     });
 
-    it.skip('renders add test plan modal on button click', async () => {
+    it('does not render ManageTestQueue component', async () => {
       // allow page time to load
       await waitFor(() => new Promise(res => setTimeout(res, 0)));
 
-      const { queryByTestId } = wrapper;
-      const button = queryByTestId('test-queue-add-test-plan-to-queue-button');
-
-      expect(button).toBeTruthy();
-
-      // opens modal
-      fireEvent.click(button);
-
-      expect(screen.getByText('Select an AT and Version')).toBeInTheDocument();
+      const { queryByText } = wrapper;
       expect(
-        screen.getByText('Select a Browser and Version')
-      ).toBeInTheDocument();
+        queryByText('Manage Assistive Technology Versions')
+      ).not.toBeInTheDocument();
       expect(
-        screen.getByText('Select a Test Plan and Version')
-      ).toBeInTheDocument();
+        queryByText('Add Test Plans to the Test Queue')
+      ).not.toBeInTheDocument();
     });
   });
 
-  describe('[IS ADMIN] when test plan reports exist', () => {
+  describe('is admin when no test plan reports exist', () => {
     beforeEach(() => {
-      wrapper = setup(TEST_QUEUE_PAGE_ADMIN_POPULATED_MOCK_DATA);
+      wrapper = setup(TEST_QUEUE_PAGE_ADMIN_NOT_POPULATED_MOCK_DATA);
     });
 
     it('renders loading state on initialization', async () => {
-      const { getByTestId } = wrapper;
-      const element = getByTestId('page-status');
+      const { queryByTestId } = wrapper;
+      const element = queryByTestId('page-status');
 
       expect(element).toBeTruthy();
       expect(element).toHaveTextContent('Loading');
     });
 
-    it.skip('renders Test Queue page instructions', async () => {
+    it('renders Test Queue page instructions', async () => {
       // allow page time to load
       await waitFor(() => new Promise(res => setTimeout(res, 0)));
 
-      const { queryByTestId, getByTestId } = wrapper;
+      const { queryByTestId } = wrapper;
       const loadingElement = queryByTestId('page-status');
-      const element = getByTestId('test-queue-instructions');
+      const noTestPlansMessageElement = queryByTestId('no-test-plans');
+      const addTestPlanToQueueMessageElement = queryByTestId(
+        'add-test-plans-queue'
+      );
 
       expect(loadingElement).not.toBeInTheDocument();
-      expect(element).toBeTruthy();
-      expect(element).toHaveTextContent(
-        'Assign yourself a test plan or start executing one that is already assigned to you.'
+      expect(noTestPlansMessageElement).toBeTruthy();
+      expect(noTestPlansMessageElement).toHaveTextContent(
+        /There are currently no test plan reports available/i
+      );
+      expect(addTestPlanToQueueMessageElement).toHaveTextContent(
+        /Add a Test Plan to the Queue/i
       );
     });
 
-    it('renders AT-specific sections', async () => {
+    it('renders no AT-specific sections', async () => {
       // allow page time to load
       await waitFor(() => new Promise(res => setTimeout(res, 0)));
 
@@ -265,46 +141,49 @@ describe('Render TestQueue/index.jsx', () => {
       const jawsElements = queryAllByText(/jaws/i);
       const voiceOverElements = queryAllByText(/voiceover/i);
 
-      expect(nvdaElements.length).toBeGreaterThanOrEqual(1);
-      expect(jawsElements.length).toBeGreaterThanOrEqual(1);
-      expect(voiceOverElements.length).toBeGreaterThanOrEqual(1);
+      expect(nvdaElements.length).toEqual(0);
+      expect(jawsElements.length).toEqual(0);
+      expect(voiceOverElements.length).toEqual(0);
     });
 
-    it('renders testers are assigned to Test Plans', async () => {
+    it('renders ManageTestQueue component', async () => {
       // allow page time to load
       await waitFor(() => new Promise(res => setTimeout(res, 0)));
 
-      const { queryAllByText } = wrapper;
-      const userAAssignedElements = queryAllByText(/esmeralda-baggins/i);
-      const userBAssignedElements = queryAllByText(/tom-proudfeet/i);
-      const assignedTestsElements = queryAllByText(
-        /\d+ of \d+ tests complete/i
-      );
-
-      expect(userAAssignedElements.length).toBeGreaterThanOrEqual(1);
-      expect(userBAssignedElements.length).toBeGreaterThanOrEqual(1);
-      expect(assignedTestsElements.length).toBeGreaterThanOrEqual(1);
+      const { queryByText } = wrapper;
+      expect(
+        queryByText('Manage Assistive Technology Versions')
+      ).toBeInTheDocument();
+      expect(
+        queryByText('Add Test Plans to the Test Queue')
+      ).toBeInTheDocument();
     });
 
-    it.skip('renders add test plan modal on button click', async () => {
+    it('renders add test plan modal on button click', async () => {
       // allow page time to load
       await waitFor(() => new Promise(res => setTimeout(res, 0)));
 
-      const { queryByTestId } = wrapper;
-      const button = queryByTestId('test-queue-add-test-plan-to-queue-button');
+      const { queryByRole, queryByText } = wrapper;
+      const manageAssistiveTechnologyVersionsButton = queryByRole('button', {
+        name: 'Manage Assistive Technology Versions'
+      });
+      const addTestPlansToTestQueueButton = queryByRole('button', {
+        name: 'Add Test Plans to the Test Queue'
+      });
 
-      expect(button).toBeTruthy();
+      fireEvent.click(manageAssistiveTechnologyVersionsButton);
+      fireEvent.click(addTestPlansToTestQueueButton);
 
-      // opens modal
-      fireEvent.click(button);
-
-      expect(screen.getByText('Select an AT and Version')).toBeInTheDocument();
       expect(
-        screen.getByText('Select a Browser and Version')
-      ).toBeInTheDocument();
+        queryByText(
+          'Select an assistive technology and manage its versions in the ARIA-AT App'
+        )
+      ).toBeVisible();
       expect(
-        screen.getByText('Select a Test Plan and Version')
-      ).toBeInTheDocument();
+        queryByText(
+          'Select a test plan, assistive technology and browser to add a new test plan report to the test queue.'
+        )
+      ).toBeVisible();
     });
   });
 });
