+++ client/tests/__mocks__/GraphQLMocks/TestQueuePageTesterNotPopulatedMock.js
@@ -6,42 +6,41 @@ export default testQueuePageQuery => [
     result: {
       data: {
         me: {
+          __typename: 'User',
           id: '4',
           username: 'bar-foo',
-          roles: ['TESTER'],
-          __typename: 'User'
+          roles: ['TESTER']
         },
-        ats: [],
-        browsers: [],
         users: [
           {
+            __typename: 'User',
             id: '1',
             username: 'foo-bar',
             roles: ['ADMIN', 'TESTER'],
             isBot: false,
-            ats: [],
-            __typename: 'User'
+            ats: []
           },
           {
+            __typename: 'User',
             id: '4',
             username: 'bar-foo',
             roles: ['TESTER'],
             isBot: false,
-            ats: [],
-            __typename: 'User'
+            ats: []
           },
           {
+            __typename: 'User',
             id: '5',
             username: 'boo-far',
             roles: ['TESTER'],
             isBot: false,
-            ats: [],
-            __typename: 'User'
+            ats: []
           }
         ],
+        ats: [],
+        testPlans: [],
         testPlanVersions: [],
-        testPlanReports: [],
-        testPlans: []
+        testPlanReports: []
       }
     }
   }
